+++ pkg/notifiers/mocks/metadata_getter.go
@@ -48,3 +48,17 @@ func (mr *MockMetadataGetterMockRecorder) GetAnnotationValue(arg0, arg1, arg2, a
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetAnnotationValue", reflect.TypeOf((*MockMetadataGetter)(nil).GetAnnotationValue), arg0, arg1, arg2, arg3)
 }
+
+// GetNamespaceLabels mocks base method.
+func (m *MockMetadataGetter) GetNamespaceLabels(arg0 context.Context, arg1 *storage.Alert) map[string]string {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "GetNamespaceLabels", arg0, arg1)
+	ret0, _ := ret[0].(map[string]string)
+	return ret0
+}
+
+// GetNamespaceLabels indicates an expected call of GetNamespaceLabels.
+func (mr *MockMetadataGetterMockRecorder) GetNamespaceLabels(arg0, arg1 interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetNamespaceLabels", reflect.TypeOf((*MockMetadataGetter)(nil).GetNamespaceLabels), arg0, arg1)
+}